chore: move client functions to dedicated module
commit-id:7ca0f454
